+++ src/cucu/utils.py
@@ -225,7 +225,7 @@ def get_step_image_dir(step_index, step_name):
 
 def take_saw_element_screenshot(ctx, thing, name, index, element=None):
     observed = "saw" if element else "did not see"
-    prefix = "" if index == 0 else f"{humanize.ordinal(index)} "
+    prefix = "" if index == 0 else f"{humanize.ordinal(index + 1)} "
 
     take_screenshot(
         ctx,