+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project closely adheres to [Semantic Versioning](https://semver.org/spe
 - Change - reorder core.py imports for BEHAVE_STRIP_STEPS_WITH_TRAILING_COLON
 - Change - from RuntimeError to AssertionError for step failure
 - Change - some ordering between environment.py and rundb.py
+- Fix - behave renamed "AssertionError" to "ASSERT FAILED"
 - Fix - behave renamed Status errored to error
 - Fix - consider undefined as failure of test