+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.3.28
-- Fix - option from nth dynamic dropdown 
+- Fix - option from nth dynamic dropdown
 
 ## 1.3.27
 - Add - step to open a new tab