+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project closely adheres to [Semantic Versioning](https://semver.org/spe
 
 ## 1.3.28
 - Fix - option from nth dynamic dropdown
+- Fix - report correct nth number
 
 ## 1.3.27
 - Add - step to open a new tab